requirements.txt
Pillow==10.0.0
PyInstaller==5.13.0
pystray==0.19.4
PyYAML==6.0.1
requests==2.28.2
